+++ sdk/src/wallet/storage/manager.rs
@@ -8,10 +8,13 @@ use serde::{Deserialize, Serialize};
 use tokio::sync::RwLock;
 
 use crate::{
-    client::secret::{SecretManager, SecretManagerDto},
+    client::{
+        secret::{SecretManager, SecretManagerDto},
+        storage::StorageAdapter as ClientStorageAdapter,
+    },
     wallet::{
         account::{AccountDetails, SyncOptions},
-        migration::migrate_storage,
+        migration::migrate,
         storage::{constants::*, Storage, StorageAdapter},
         WalletBuilder,
     },
@@ -51,14 +54,14 @@ pub struct StorageManager {
 
 impl StorageManager {
     pub(crate) async fn new(
-        storage: impl StorageAdapter + Send + Sync + 'static,
+        storage: impl StorageAdapter + 'static,
         encryption_key: impl Into<Option<[u8; 32]>> + Send,
     ) -> crate::wallet::Result<Self> {
         let storage = Storage {
             inner: Box::new(storage) as _,
             encryption_key: encryption_key.into(),
         };
-        migrate_storage(&storage).await?;
+        migrate(&storage).await?;
 
         // Get the db version or set it
         if let Some(db_schema_version) = storage.get::<u8>(DATABASE_SCHEMA_VERSION_KEY).await? {
@@ -69,7 +72,7 @@ impl StorageManager {
             }
         } else {
             storage
-                .set(DATABASE_SCHEMA_VERSION_KEY, DATABASE_SCHEMA_VERSION)
+                .set(DATABASE_SCHEMA_VERSION_KEY, &DATABASE_SCHEMA_VERSION)
                 .await?;
         };
 
@@ -108,7 +111,7 @@ impl StorageManager {
             match secret_manager_dto {
                 SecretManagerDto::Mnemonic(_) => {}
                 _ => {
-                    self.storage.set(SECRET_MANAGER_KEY, secret_manager_dto).await?;
+                    self.storage.set(SECRET_MANAGER_KEY, &secret_manager_dto).await?;
                 }
             }
         }
@@ -164,9 +167,7 @@ impl StorageManager {
             self.account_indexes.push(*account.index());
         }
 
-        self.storage
-            .set(ACCOUNTS_INDEXATION_KEY, self.account_indexes.clone())
-            .await?;
+        self.storage.set(ACCOUNTS_INDEXATION_KEY, &self.account_indexes).await?;
         self.storage
             .set(&format!("{ACCOUNT_INDEXATION_KEY}{}", account.index()), account)
             .await
@@ -174,12 +175,10 @@ impl StorageManager {
 
     pub async fn remove_account(&mut self, account_index: u32) -> crate::wallet::Result<()> {
         self.storage
-            .remove(&format!("{ACCOUNT_INDEXATION_KEY}{account_index}"))
+            .delete(&format!("{ACCOUNT_INDEXATION_KEY}{account_index}"))
             .await?;
         self.account_indexes.retain(|a| a != &account_index);
-        self.storage
-            .set(ACCOUNTS_INDEXATION_KEY, self.account_indexes.clone())
-            .await
+        self.storage.set(ACCOUNTS_INDEXATION_KEY, &self.account_indexes).await
     }
 
     pub async fn set_default_sync_options(
@@ -188,7 +187,7 @@ impl StorageManager {
         sync_options: &SyncOptions,
     ) -> crate::wallet::Result<()> {
         let key = format!("{ACCOUNT_INDEXATION_KEY}{account_index}-{ACCOUNT_SYNC_OPTIONS}");
-        self.storage.set(&key, sync_options.clone()).await
+        self.storage.set(&key, &sync_options).await
     }
 
     pub async fn get_default_sync_options(&self, account_index: u32) -> crate::wallet::Result<Option<SyncOptions>> {
@@ -200,12 +199,12 @@ impl StorageManager {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::wallet::storage::adapter::memory::{Memory, STORAGE_ID};
+    use crate::{client::storage::StorageAdapterId, wallet::storage::adapter::memory::Memory};
 
     #[tokio::test]
     async fn id() {
         let storage_manager = StorageManager::new(Memory::default(), None).await.unwrap();
-        assert_eq!(storage_manager.id(), STORAGE_ID);
+        assert_eq!(storage_manager.id(), Memory::ID);
         assert!(!storage_manager.is_encrypted());
     }
 
@@ -224,7 +223,7 @@ mod tests {
             c: -420,
         };
         let storage = Memory::default();
-        storage.set("key", serde_json::to_string(&rec).unwrap()).await.unwrap();
+        storage.set("key", &rec).await.unwrap();
 
         let storage_manager = StorageManager::new(storage, None).await.unwrap();
         assert_eq!(Some(rec), storage_manager.get::<Record>("key").await.unwrap());