+++ sdk/CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Stronghold snapshot migration from v2 to v3;
 - `SecretManage::sign_evm`;
 - `Account::addresses_balance` method accepting addresses to get balance for;
+- `Wallet::get_secret_manager` method;
 
 ### Changed
 
@@ -73,6 +74,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `AccountInner::incoming_transactions` returns a `Vec` instead of a `HashMap`;
 - `Address::try_from_bech32_with_hrp` refactored to `try_from_bech32`;
 - `{MetadataFeature, TagFeature}::new` take an `impl Into<Vec<u8>>` param;
+- Merged `StorageProvider` into `StorageAdapter`;
 - `GetAddressesBuilderOptions` renamed to `GetAddressesOptions` and fields no longer nullable;
 - Methods on `GetAddressesBuilder` moved to `SecretManager`;
 - Rename `GenerateAddresses` method to `GenerateEd25519Addresses` for Account and SecretManager, and their respective responses;